ISSUE-345: fixtures in folders
