+++ tests/Integration/Controller/AbstractTestController.php
@@ -6,8 +6,8 @@ namespace PhpList\RestBundle\Tests\Integration\Controller;
 
 use Doctrine\ORM\Tools\SchemaTool;
 use PhpList\Core\TestingSupport\Traits\DatabaseTestTrait;
-use PhpList\RestBundle\Tests\Integration\Controller\Fixtures\AdministratorFixture;
-use PhpList\RestBundle\Tests\Integration\Controller\Fixtures\AdministratorTokenFixture;
+use PhpList\RestBundle\Tests\Integration\Controller\Fixtures\Identity\AdministratorFixture;
+use PhpList\RestBundle\Tests\Integration\Controller\Fixtures\Identity\AdministratorTokenFixture;
 use Symfony\Bundle\FrameworkBundle\Test\WebTestCase;
 use Symfony\Component\DomCrawler\Crawler;
 use Symfony\Component\HttpFoundation\Response;
